Refactor calculations to work with doubles

+++ core/src/main/scala/currexx/core/signal/SignalService.scala
@@ -67,7 +67,7 @@ object SignalService:
 
   def detectMacd(uid: UserId, data: MarketTimeSeriesData, macd: IndicatorParameters.MACD): Option[Signal] = {
     val (macdLine, signalLine) = MovingAverages.macdWithSignal(
-      values = data.prices.map(_.close).toList,
+      values = data.prices.map(_.close.toDouble).toList,
       fastLength = macd.fastLength,
       slowLength = macd.slowLength,
       signalSmoothing = macd.signalSmoothing
@@ -78,7 +78,7 @@ object SignalService:
   }
 
   def detectRsi(uid: UserId, data: MarketTimeSeriesData, rsi: IndicatorParameters.RSI): Option[Signal] = {
-    val rsis    = MomentumOscillators.rsi(data.prices.map(_.close).toList, rsi.length)
+    val rsis    = MomentumOscillators.rsi(data.prices.map(_.close.toDouble).toList, rsi.length)
     val currRsi = rsis.head
     def above   = Option.when(currRsi > rsi.upperLine)(Condition.AboveThreshold(BigDecimal(rsi.upperLine), currRsi))
     def below   = Option.when(currRsi < rsi.lowerLine)(Condition.BelowThreshold(BigDecimal(rsi.lowerLine), currRsi))
@@ -89,9 +89,9 @@ object SignalService:
 
   def detectStoch(uid: UserId, data: MarketTimeSeriesData, stoch: IndicatorParameters.STOCH): Option[Signal] = {
     val (k, d) = MomentumOscillators.stoch(
-      closings = data.prices.map(_.close).toList,
-      highs = data.prices.map(_.high).toList,
-      lows = data.prices.map(_.low).toList,
+      closings = data.prices.map(_.close.toDouble).toList,
+      highs = data.prices.map(_.high.toDouble).toList,
+      lows = data.prices.map(_.low.toDouble).toList,
       length = stoch.length,
       slowKLength = stoch.slowKLength,
       slowDLength = stoch.slowDLength
@@ -105,7 +105,7 @@ object SignalService:
   }
 
   def detectHma(uid: UserId, data: MarketTimeSeriesData, hma: IndicatorParameters.HMA): Option[Signal] = {
-    val hmas  = MovingAverages.hma(data.prices.toList.map(_.close), hma.length).toArray.take(5)
+    val hmas  = MovingAverages.hma(data.prices.toList.map(_.close.toDouble), hma.length).toArray.take(5)
     val hmas2 = hmas.tail
     val hmas3 = hmas.zip(hmas.map(-_)).map(_ - _)
     val hmas4 = hmas.zip(hmas).map(_ + _)
